worked on logics of customer and driver

+++ TaxiBookingService.Logic/User/AdminLogic.cs
@@ -1,18 +1,9 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
-using Microsoft.IdentityModel.Tokens;
-using System;
-using System.Collections.Generic;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Security.Cryptography;
-using System.Text;
-using System.Threading.Tasks;
 using TaxiBookingService.API.User.Admin;
-using TaxiBookingService.Common;
 using TaxiBookingService.Common.AssetManagement.Common;
-//using TaxiBookingService.Common.Enums;
+using TaxiBookingService.Common.Utilities;
 using TaxiBookingService.Dal.Entities;
 using TaxiBookingService.Dal.Interfaces;
 using TaxiBookingService.Logic.User.Interfaces;
@@ -20,70 +11,21 @@ using static TaxiBookingService.Common.CustomException;
 
 namespace TaxiBookingService.Logic.User
 {
-    public class AdminLogic : IAdminLogic<Admin>
+    public class AdminLogic : IAdminLogic
     {
         private readonly IConfiguration _configuration;
         private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
-        public AdminLogic(IUnitOfWork unitOfWork, IConfiguration configuration, IHttpContextAccessor httpContextAccessor,IMapper mapper)
+        private readonly ILoggerAdapter _loggerAdapter;
+
+        public AdminLogic(IUnitOfWork unitOfWork, IConfiguration configuration, IHttpContextAccessor httpContextAccessor,IMapper mapper, ILoggerAdapter loggerAdapter)
         {
             _unitOfWork = unitOfWork;
             _configuration = configuration;
             _httpContextAccessor = httpContextAccessor;
             _mapper=mapper;
-        }
-
-        private string CreateToken(Admin admin)
-        {
-            var claims = new List<Claim>
-            {
-                new Claim(ClaimTypes.Email, admin.User.Email),
-                new Claim(ClaimTypes.Role, "Admin"), 
-
-            };
-
-            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration.GetSection("AppSettings:Token").Value));
-            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha512Signature);
-
-            var token = new JwtSecurityToken(
-                claims: claims,
-                expires: DateTime.Now.AddMinutes(15),
-                signingCredentials: creds
-            );
-
-            return new JwtSecurityTokenHandler().WriteToken(token);
-        }
-
-        private bool VerifyPasswordHash(string password, byte[] passwordHash, byte[] passwordSalt)
-        {
-            using (var hmac = new HMACSHA512(passwordSalt))
-            {
-                var computedHash = hmac.ComputeHash(Encoding.UTF8.GetBytes(password));
-                return computedHash.SequenceEqual(passwordHash);
-            }
-        }
-
-        public async Task<string> Login(AdminLoginDto request)
-        {
-            var admin = await _unitOfWork.AdminRepository.GetByEmail(request.Email);
-            if (admin == null)
-            {
-                throw new Exception("Invalid email or password.");
-            }
-            if (!VerifyPasswordHash(request.Password, admin.User.PasswordHash, admin.User.PasswordSalt))
-            {
-                throw new AuthenticationException(AppConstant.PasswordNotFound);
-            }
-            string token = CreateToken(admin);
-            _httpContextAccessor.HttpContext.Response.Cookies.Append("accessToken", token);
-            await _unitOfWork.SaveChangesAsync();
-            return token;
-        }
-
-        public async Task Logout()
-        {
-          _httpContextAccessor.HttpContext.Response.Cookies.Delete("accessToken");
+            _loggerAdapter = loggerAdapter;
         }
 
         public async Task<int> AddCancellationReason(AdminManageReasonDto request)
@@ -96,24 +38,41 @@ namespace TaxiBookingService.Logic.User
 
         public async Task<bool> DeleteCancellationReason(int Id)
         {
-            var exisitingreason=await _unitOfWork.RideCancellationReasonRepository.GetById(Id);
-            await _unitOfWork.RideCancellationReasonRepository.Delete(exisitingreason);
+            var existingReason = await _unitOfWork.RideCancellationReasonRepository.GetById(Id);
+
+            if (existingReason == null)
+            {
+                throw new NotFoundException(AppConstant.ReasonNotFound, _loggerAdapter);
+            }
+
+            await _unitOfWork.RideCancellationReasonRepository.Delete(existingReason);
             await _unitOfWork.SaveChangesAsync();
             return true;
         }
 
         public async Task<int> UpdateCancellationReason(AdminManageReasonDto request,int Id)
         {
-            var exisitingreason = await _unitOfWork.RideCancellationReasonRepository.GetById(Id);
-            _mapper.Map(request, exisitingreason);
-            await _unitOfWork.RideCancellationReasonRepository.Update(exisitingreason);
+            var existingReason = await _unitOfWork.RideCancellationReasonRepository.GetById(Id);
+
+            if (existingReason == null)
+            {
+                throw new NotFoundException(AppConstant.ReasonNotFound, _loggerAdapter);
+            }
+
+            _mapper.Map(request, existingReason);
+            await _unitOfWork.RideCancellationReasonRepository.Update(existingReason);
             await _unitOfWork.SaveChangesAsync();
-            return exisitingreason.Id;
+            return existingReason.Id;
         }
 
         public async Task<bool> DeleteUser(int Id)
         {
             var exisitingUser = await _unitOfWork.UserRepository.GetById(Id);
+
+            if (exisitingUser == null) { 
+                throw new NotFoundException(AppConstant.UserNotFound,_loggerAdapter); 
+            }
+
             await _unitOfWork.UserRepository.Delete(exisitingUser);
             await _unitOfWork.SaveChangesAsync();
             return true;
@@ -122,6 +81,11 @@ namespace TaxiBookingService.Logic.User
         public async Task<int> UpdateUser(AdminManageUserDto request, int Id)
         {
             var exisitingUser = await _unitOfWork.UserRepository.GetById(Id);
+
+            if (exisitingUser == null) { 
+                throw new NotFoundException(AppConstant.UserNotFound,_loggerAdapter); 
+            }
+
             _mapper.Map(request, exisitingUser);
             await _unitOfWork.UserRepository.Update(exisitingUser);
             await _unitOfWork.SaveChangesAsync();
